Repackage contact- and event-related parsers
