+++ src/test/java/seedu/address/logic/parser/AddContactCommandParserTest.java
@@ -28,6 +28,7 @@ import static seedu.address.testutil.TypicalPersons.BOB;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.contacts.AddContactCommand;
+import seedu.address.logic.parser.contacts.AddContactCommandParser;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
